More addons
- Water turbulence finished
- Engine can be build now against x64 architecture
- Fixed few compilation warnings
- Abstract window system, first phase

+++ RenderEngine/RenderEngine/include/DeferredRenderObject.h
@@ -40,7 +40,7 @@ namespace Engine
 
 		unsigned int getFrameBufferId();
 
-		TextureInstance * addColorBuffer(unsigned int index, GLenum gpuTextureFormat, GLenum inputTextureFormat, GLenum pixelFormat, unsigned int w, unsigned int h, float filterMethod = GL_NEAREST);
+		TextureInstance * addColorBuffer(unsigned int index, GLenum gpuTextureFormat, GLenum inputTextureFormat, GLenum pixelFormat, unsigned int w, unsigned int h, int filterMethod = GL_NEAREST);
 		TextureInstance * addDepthBuffer24(unsigned int w, unsigned int h);
 		TextureInstance * addDepthBuffer32(unsigned int w, unsigned int h);
